requirements.txt
arrow==0.8.0
coreapi==2.0.9
Django==1.10.3
django-cors-headers==1.1.0
django-dbbackup==3.0.3
django-filter==0.15.3
django-rest-swagger==2.1.0
djangorestframework==3.5.3
djangorestframework-filters==0.8.1
djangorestframework-jwt==1.8.0
itypes==1.1.0
MySQL-python==1.2.5
openapi-codec==1.1.7
PyJWT==1.4.2
python-dateutil==2.6.0
pytz==2016.7
redis==2.10.5
requests==2.11.1
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
